Adding scheduled actions enable/disable
Why? Because due to various CF/ebextensions bugs doing conditional
scheduled actions is difficult. As an interim we're applying outside of
CF - however this can break the deployment if an instance is phoenix'd
because the ASG has a scheduled action that CF doesn't recognise. This
is a workaround until we finally nail conditonal scheduled actions in
ebextenisons.